# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: 'SAKA Marketplace API'
  version: 1.0.0
  description: |-
    Multi-vertical marketplace API for Tanzania.

    **Errors.** Every failure returns the same envelope:

        { "error": { "code": "...", "message": "...", "details": {...}, "request_id": "..." } }

    Clients should switch on `error.code`, never on `message`. Validation failures (422) additionally carry Laravel's `errors` map.

    **Correlation.** Every response carries `X-Request-Id`; an inbound value is honoured when it matches `^[A-Za-z0-9_-]{8,64}$`.

    **Authentication.** Bearer tokens issued by `/auth/login`. Tokens are unscoped; authorization is enforced by server-side policies.

    **Not found vs forbidden.** Resources the caller may not see return 404, never 403, so existence is not disclosed.
  contact:
    name: 'SAKA Engineering'
  license:
    name: Proprietary
servers:
  -
    url: 'https://api.saka.co.tz/api/v1'
    description: Production
  -
    url: 'http://localhost:8000/api/v1'
    description: Local
...
paths:
...
  /auth/forgot-password:
    post:
      operationId: api_v1_auth_password_forgot
      summary: 'Request a password reset link'
      tags:
        - Authentication
      responses:
        202:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Message'
        422:
          $ref: '#/components/responses/ValidationFailed'
        429:
          $ref: '#/components/responses/RateLimited'
        500:
          $ref: '#/components/responses/ServerError'
      description: |-
        Always returns 202 whether or not the address is registered.

        Rate limit: `password-reset`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
  /auth/login:
    post:
      operationId: api_v1_auth_login
      summary: 'Sign in'
      tags:
        - Authentication
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AuthResult'
        401:
          $ref: '#/components/responses/InvalidCredentials'
        403:
          $ref: '#/components/responses/AccountBlocked'
        422:
          $ref: '#/components/responses/ValidationFailed'
        429:
          $ref: '#/components/responses/RateLimited'
        500:
          $ref: '#/components/responses/ServerError'
      description: |-
        An unknown email and a wrong password return byte-identical responses to prevent account enumeration.

        Rate limit: `auth-login`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
...
  /auth/me:
    get:
      operationId: api_v1_auth_me
      summary: 'Current user'
      tags:
        - Authentication
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        429:
          $ref: '#/components/responses/RateLimited'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        -
          bearerAuth: []
...
  /auth/register:
    post:
      operationId: api_v1_auth_register
      summary: 'Register with email and password'
      tags:
        - Authentication
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AuthResult'
        422:
          $ref: '#/components/responses/ValidationFailed'
        429:
          $ref: '#/components/responses/RateLimited'
        500:
          $ref: '#/components/responses/ServerError'
      description: |-


        Rate limit: `auth-register`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
  /auth/reset-password:
    post:
      operationId: api_v1_auth_password_reset
      summary: 'Reset a password with a token'
      tags:
        - Authentication
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Message'
        422:
          $ref: '#/components/responses/ValidationFailed'
        429:
          $ref: '#/components/responses/RateLimited'
        500:
          $ref: '#/components/responses/ServerError'
      description: |-
        Revokes every existing token for the account.

        Rate limit: `password-reset`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Token from `/auth/login`, `/auth/register` or `/auth/oauth/google`.'
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
          required:
            - code
            - message
      required:
        - error
    ValidationError:
      allOf:
        -
          $ref: '#/components/schemas/Error'
        -
          type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
    Message:
      type: object
      properties:
        message:
          type: string
...
    User:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type:
            - string
            - 'null'
        full_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type:
            - string
            - 'null'
        locale:
          type: string
        status:
          type: string
          enum:
            - pending
            - active
            - suspended
            - banned
        email_verified:
          type: boolean
        phone_verified:
          type: boolean
        can_publish_listings:
          type: boolean
        avatar_url:
          type:
            - string
            - 'null'
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    AuthResult:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
        token_type:
          type: string
          enum:
            - Bearer
        expires_at:
          type: string
          format: date-time
...
    RegisterRequest:
      type: object
      properties:
        first_name:
          type: string
          minLength: 2
          maxLength: 100
        last_name:
          type:
            - string
            - 'null'
        email:
          type: string
          format: email
        phone:
          type:
            - string
            - 'null'
        password:
          type: string
          minLength: 8
        password_confirmation:
          type: string
        device_name:
          type: string
      required:
        - first_name
        - email
        - password
        - password_confirmation
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        device_name:
          type: string
      required:
        - email
        - password
...
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
        email:
          type: string
        password:
          type: string
        password_confirmation:
          type: string
      required:
        - token
        - email
        - password
        - password_confirmation
...
  responses:
    Unauthenticated:
      description: 'Missing or invalid token (`UNAUTHENTICATED`).'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 'Authenticated but not permitted (`FORBIDDEN`).'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InvalidCredentials:
      description: 'Credentials rejected (`INVALID_CREDENTIALS`). Identical for an unknown account and a wrong password.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    AccountBlocked:
      description: 'Account suspended or banned (`ACCOUNT_SUSPENDED`, `ACCOUNT_BANNED`).'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationFailed:
      description: 'Validation failed (`VALIDATION_FAILED`). Carries an `errors` map.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    RateLimited:
      description: 'Rate limit exceeded (`RATE_LIMITED`). See `Retry-After`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: 'Unexpected error (`SERVER_ERROR`). No internal detail is disclosed.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'